Break out previewer actions.
